Move status_type_enum files
